Port the code (and disable Fabric) to the zend_hash compatibility layer
The code compiles but is not correct due to the semantics of pre-PHP7
where a double pointer needed to be passed to the hash functions while
PHP7+ expects just a pointer (to the _ptr) functions which is stored
a copy of the pointer and not the referenced memory.

The next step is to move the code to the semantics of PHP7. This will
break compatibility with pre-PHP7. In other words, mysqlnd_ms 1.7+ will
be only PHP7 ready. For now it compiles with 7.0 and won't compile with 7.1
due to changes in mysqlnd. This will be fixed at later stage with some macro
magic.